Moved ItemCombinations.cs to a new folder.
