+++ ItemCombinations/Program.cs
@@ -2,6 +2,7 @@
 using System.Text.Json;
 using System.Text.Json.Serialization;
 using UWRandomizerEditor;
+using UWRandomizerEditor.CMBdotDAT;
 
 namespace ItemCombinations;
